requirements.txt
Django==5.2.10
djangorestframework
djoser
django-cors-headers
django-extensions
python-dotenv
django-storages
boto3
Pillow
firebase-admin
